+++ CHANGELOG.md
@@ -2905,3 +2905,4 @@
 - [2026-07-16 19:57] feat(network): add bi-directional event bus handler for network
 - [2026-07-16 20:01] feat(cache): implement responsive cache view component
 - [2026-07-16 21:12] feat(i18n): implement shimmer loading skeleton in i18n
+- [2026-07-16 21:13] test(ui): add unit tests for ui business logic